test(validator-*): move to uvu
