+++ packages/react/tsconfig.json
@@ -13,5 +13,5 @@
     "jsx": "react",
     "declarationDir": "./dist/types"
   },
-  "include": ["src"]
+  "include": ["src", "tests"]
 }
