+++ packages/reporter-react/tsconfig.json
@@ -10,6 +10,6 @@
     "outDir": "./dist",
     "module": "esnext"
   },
-  "include": ["./src"],
+  "include": ["./src", "tests"],
   "exclude": ["node_modules/"]
 }
