--- packages/validator-superstruct/jest.config.js
@@ -1,5 +1,0 @@
-export default {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  collectCoverageFrom: ['./src/**'],
-};
